feat: Extend TrapManager constructor with the optional currentActivity
It is needed in cases when the TrapManager is instantiated after the
first activity was created. In this case the subscription for the
window callbacks of this activity is done in the TrapManager
constructor. Because the TrapManager is not notified about the
onActivityCreated event for this activity, as it subscribed for the
application lifecycle events after this activity was created.